♻️ refactor: type -\> constant로 변경
