--- server/src/chat/chat.type.ts
@@ -1,7 +1,0 @@
-export type BroadcastPayload = {
-  messageId: string;
-  userId: string;
-  username: string;
-  message: string;
-  timestamp: Date;
-};
